Refactor Parser + Amiibo Suit bugfix
Fixed a bug where the DX Amiibo Suit (and all suits after it) had the wrong ID.

Refactored the parser, splitting it up into a parser and a serializer. Got rid of the weird filename pattern handling.
Removed the UnicodeHelper as the format turned out to be big endian (opposed to little endian as assumed earlier)