Better Logging + command line arguments + bugfixes
Set up a default file logger, allowing exceptions to be logged even in PyInstaller's windowed mode. Exceptions are logged to error.log by default.

Added a .env file for logger settings.

Allow ghost files to be loaded when "dragged onto" Poltergust in the file explorer, or when a ghost file is passed as a command line argument.

Fixed a bug where the plus and magnifying glass icons of the CT manager were loaded incorrectly.